03-api-solid/src/use-cases/get-user-profile.ts
import { User } from "@prisma/client";
import { ResourceNotFoundError } from "errors/resourcer-not-found-error";
import { IUsersRepository } from "repositories/users-repository";

interface IGetUserProfileCaseRequest {
  userId: string;
}

interface IGetUserProfileUseCaseResponse {
  user: User;
}

export class GetUserProfileUseCase {
  constructor(private usersRepository: IUsersRepository) {}

  async execute({
    userId,
  }: IGetUserProfileCaseRequest): Promise<IGetUserProfileUseCaseResponse> {
    const user = await this.usersRepository.findById(userId);

    if (!user) {
      throw new ResourceNotFoundError();
    }

    return { user };
  }
}
